fix some stuff maybe
