+++ src/main/java/io/github/moulberry/notenoughupdates/listener/ItemTooltipListener.java
@@ -789,6 +789,7 @@ public class ItemTooltipListener {
 						}
 					}
 
+					System.out.println(event.toolTip);
 					PetInfoOverlay.Pet pet = PetInfoOverlay.getPetFromStack(
 						event.itemStack.getTagCompound()
 					);
